Refactor frontend architecture and fix critical loading bug
This major refactoring addresses the core architectural issues of the application to make it functional and production-ready.

Key changes include:
- **Fixed Critical Loading Bug:** The application was previously unusable as it was stuck on a loading screen. This was caused by a `ReferenceError` due to improper global state management.
- **Modularized JavaScript:** The monolithic `script.js` (nearly 7,000 lines) has been broken down into modern ES6 modules.
  - `main.js`: New entry point for the application.
  - `api.js`: A new mock API module that separates the hardcoded location data from the application logic.
  - `state.js`: A new state management module to handle shared state like user preferences and favorites, resolving the `ReferenceError`.
- **Corrected Project Structure:**
  - Removed redundant `index.html`, `script.js`, and `styles.css` from the root directory.
  - Promoted the correct `index.html` from `src/pages/home` to the root.
  - Fixed all broken asset paths in `index.html`.
- **Enabled Development Server:** The `package.json` was cleaned up, and the application is now runnable via a simple Python HTTP server.

This foundational work resolves the most critical bugs and establishes a maintainable and scalable architecture, paving the way for future UX/UI enhancements.
