+++ package.json
@@ -2,7 +2,6 @@
   "name": "painaidee-map-3d",
   "version": "1.0.0",
   "description": "PaiNaiDee 3D Interactive Map - แผนที่ 3 มิติแบบอินเทอแรคทีฟสำหรับการท่องเที่ยวประเทศไทย",
-  "main": "src/pages/home/index.html",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "python3 -m http.server 8080",
