--- script.js
@@ -1,1 +1,0 @@
-src/assets/js/script.js
